# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Meeting {
  id: number;
  name: string;
  /** Day of week: Monday, Tuesday, etc. */
  day: string;
  /** HH:MM 24-hour format */
  startTime: string;
  /** HH:MM 24-hour format */
  endTime: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  link?: string | null;
  /** Open, Closed, Beginners, etc. */
  type: string;
  /** Discussion, Book Study, Speaker, Call-Up */
  format: string;
  /**
     * Big Book, 12 & 12, Grapevine, etc.
     * @nullable
     */
  literature?: string | null;
  /** In-Person, Online, Hybrid */
  interaction: string;
  /** English, Spanish, etc. */
  language: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface MeetingInput {
  name: string;
  day: string;
  startTime: string;
  endTime: string;
  location?: string;
  link?: string;
  type: string;
  format: string;
  literature?: string;
  interaction: string;
  language: string;
  notes?: string;
}

export interface MeetingUpdate {
  name?: string;
  day?: string;
  startTime?: string;
  endTime?: string;
  location?: string;
  link?: string;
  type?: string;
  format?: string;
  literature?: string;
  interaction?: string;
  language?: string;
  notes?: string;
}

export interface PersonAssignment {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  assignedRole?: string | null;
}

export interface MeetingDetail {
  id: number;
  name: string;
  day: string;
  startTime: string;
  endTime: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  link?: string | null;
  type: string;
  format: string;
  /** @nullable */
  literature?: string | null;
  interaction: string;
  language: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
  people: PersonAssignment[];
}

export interface Person {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cleanDate?: string | null;
  createdAt?: string;
}

export interface PersonInput {
  name: string;
  role: string;
  phone?: string;
  email?: string;
  cleanDate?: string;
}

export interface PersonUpdate {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
  cleanDate?: string;
}

export interface PersonDetail {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cleanDate?: string | null;
  createdAt?: string;
  meetings: Meeting[];
}

export interface AssignmentInput {
  personId: number;
  assignedRole?: string;
}

export interface DaySchedule {
  day: string;
  meetings: MeetingDetail[];
}

export interface WeeklySchedule {
  days: DaySchedule[];
}

export interface CountByLabel {
  label: string;
  count: number;
}

export interface Stats {
  totalMeetings: number;
  totalPeople: number;
  meetingsByDay: CountByLabel[];
  meetingsByType: CountByLabel[];
  meetingsByInteraction: CountByLabel[];
}

export type ListMeetingsParams = {
day?: string;
type?: string;
format?: string;
interaction?: string;
language?: string;
search?: string;
};

export type ListPeopleParams = {
role?: string;
search?: string;
};