# lib/api-zod/src/generated/types/meeting.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Meeting {
  id: number;
  name: string;
  /** Day of week: Monday, Tuesday, etc. */
  day: string;
  /** HH:MM 24-hour format */
  startTime: string;
  /** HH:MM 24-hour format */
  endTime: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  link?: string | null;
  /** Open, Closed, Beginners, etc. */
  type: string;
  /** Discussion, Book Study, Speaker, Call-Up */
  format: string;
  /**
     * Big Book, 12 & 12, Grapevine, etc.
     * @nullable
     */
  literature?: string | null;
  /** In-Person, Online, Hybrid */
  interaction: string;
  /** English, Spanish, etc. */
  language: string;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}