# lib/api-zod/src/generated/types/meetingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingInput {
  name: string;
  day: string;
  startTime: string;
  endTime: string;
  location?: string;
  link?: string;
  type: string;
  format: string;
  literature?: string;
  interaction: string;
  language: string;
  notes?: string;
}