# lib/api-zod/src/generated/types/person.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Person {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cleanDate?: string | null;
  createdAt?: string;
}