# lib/api-zod/src/generated/types/personDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Meeting } from './meeting';

export interface PersonDetail {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  cleanDate?: string | null;
  createdAt?: string;
  meetings: Meeting[];
}