# lib/api-zod/src/generated/types/personInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */

export interface PersonInput {
  name: string;
  role: string;
  phone?: string;
  email?: string;
  cleanDate?: string;
}