# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AA Meetings Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CountByLabel } from './countByLabel';

export interface Stats {
  totalMeetings: number;
  totalPeople: number;
  meetingsByDay: CountByLabel[];
  meetingsByType: CountByLabel[];
  meetingsByInteraction: CountByLabel[];
}